% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legendrecoef}
\alias{legendrecoef}
\title{Coefficients for Legendre polynomials}
\format{
An 11 x 11 matrix, with each column representing one of the Legendre
polynomials P_0 through P_10, and each row giving the coefficients of one power
of x, x^0 through x^{10}.
}
\usage{
legendrecoef
}
\description{
Coefficients of the Legendre polynomials, currently given up through P_{10}.
}
\keyword{datasets}
