% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendreseries.R
\name{legendrederiv_extrema}
\alias{legendrederiv_extrema}
\title{Find the extrema of the derivative of a Legendre series}
\usage{
legendrederiv_extrema(lc)
}
\arguments{
\item{lc}{Coefficients of the Legendre series}
}
\value{
Vector with the minimum and maximum value of the derivative, in that
order
}
\description{
Find the min and max values, over the interval [-1,1] of a Legendre series,
given the coefficients of the series.
}
\details{
We find these values by finding the real zeros of the second derivative of
the series and evaluating the first derivative at those points, plus the
end points of the interval.
}
\keyword{internal}
