% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendreseries.R
\name{lpolycoef}
\alias{lpolycoef}
\title{Scale Legendre polynomials by the coefficients of a Legendre series}
\usage{
lpolycoef(lc)
}
\arguments{
\item{lc}{Coefficients of the Legendre series}
}
\description{
Scale Legendre polynomials by the coefficients of a Legendre series
}
\keyword{internal}
