% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendreseries.R
\name{lserieseval}
\alias{lserieseval}
\title{Evaluate a Legendre series at a specified set of points}
\usage{
lserieseval(x, lc, interval)
}
\arguments{
\item{x}{The values at which to evaluate the series}

\item{lc}{The coefficients of the Legendre polynomials in the series, starting
with P_0 and counting upward.  Any omitted coefficients are filled in with zeros.}

\item{interval}{Boundaries of the x interval (see details)}
}
\description{
Given the series coefficients, evaluate the Legendre series at the specified
points.
}
\details{
The Legendre polynomials are defined on the interval [-1,1].  A function defined
on a different (finite) interval can be rescaled to work with the Legendre polynomials
using the \code{interval} parameter.  If the interval is omitted, it is assumed to
be [-1,1]
}
