% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyeval.R
\name{polyeval}
\alias{polyeval}
\alias{polysumeval}
\title{Evaluate polynomials at specified x values using synthetic substitution}
\usage{
polyeval(x, a)

polysumeval(x, am)
}
\arguments{
\item{x}{Vector of values to evaluate}

\item{a}{Vector of polynomial coefficients}

\item{am}{Matrix of polynomial coefficients; each column is one polynomial}
}
\value{
Vector of P(x) values, for all x
}
\description{
Polynomials are specified by a vector of coefficients \eqn{a_n}, with \eqn{a_0} in
the first vector entry, and counting upward from there.
}
\section{Functions}{
\itemize{
\item \code{polysumeval}: Evaluate a sum of polynomials at specified values
}}

